requirements.txt
streamlit
tensorflow
Pillow
numpy
transformers
sentencepiece
torch
whisper
ffmpeg-python